pyproject.toml
[tool.poetry]
name = "blaise-cawi-load-tests"
version = "0.1.0"
description = ""
authors = ["Nik Webb <nik.webb@ons.gov.uk>"]

[tool.poetry.dependencies]
python = "^3.9"
locust = "^2.0.0"
locust-plugins = "^1.2.3"
requests = "^2.26.0"
python-dotenv = "^0.19.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
